+++ GitTest/src/main/java/com/smhrd/model/VO/GJ_CN_VO.java
@@ -5,12 +5,12 @@ public class GJ_CN_VO {
 	private int cn_seq;
 	private String cn_nm;
 	private String cn_addr;
-	private int cortarNo;
+	private long cortarNo;
 	private String dong;
-	private int lat;
-	private int lng;
+	private	double lat;
+	private double lng;
 	
-	public GJ_CN_VO(int cn_seq, String cn_nm, String cn_addr, int cortarNo, String dong, int lat, int lng) {
+	public GJ_CN_VO(int cn_seq, String cn_nm, String cn_addr, long cortarNo, String dong, double lat, double lng) {
 		this.cn_seq = cn_seq;
 		this.cn_nm = cn_nm;
 		this.cn_addr = cn_addr;
@@ -44,11 +44,11 @@ public class GJ_CN_VO {
 		this.cn_addr = cn_addr;
 	}
 
-	public int getCortarNo() {
+	public long getCortarNo() {
 		return cortarNo;
 	}
 
-	public void setCortarNo(int cortarNo) {
+	public void setCortarNo(long cortarNo) {
 		this.cortarNo = cortarNo;
 	}
 
@@ -60,19 +60,19 @@ public class GJ_CN_VO {
 		this.dong = dong;
 	}
 
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-	public int getLng() {
+	public double getLng() {
 		return lng;
 	}
 
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
