+++ GitTest/src/main/java/com/smhrd/model/VO/GJ_EX_VO.java
@@ -5,12 +5,12 @@ public class GJ_EX_VO {
 	private int ex_seq;
 	private String ex_nm;
 	private String ex_addr;
-	private int cortarNo;
+	private long cortarNo;
 	private String dong;
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
 	
-	public GJ_EX_VO(int ex_seq, String ex_nm, String ex_addr, int cortarNo, String dong, int lat, int lng) {
+	public GJ_EX_VO(int ex_seq, String ex_nm, String ex_addr, long cortarNo, String dong, double lat, double lng) {
 		this.ex_seq = ex_seq;
 		this.ex_nm = ex_nm;
 		this.ex_addr = ex_addr;
@@ -44,11 +44,11 @@ public class GJ_EX_VO {
 		this.ex_addr = ex_addr;
 	}
 
-	public int getCortarNo() {
+	public long getCortarNo() {
 		return cortarNo;
 	}
 
-	public void setCortarNo(int cortarNo) {
+	public void setCortarNo(long cortarNo) {
 		this.cortarNo = cortarNo;
 	}
 
@@ -60,21 +60,22 @@ public class GJ_EX_VO {
 		this.dong = dong;
 	}
 
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-	public int getLng() {
+	public double getLng() {
 		return lng;
 	}
 
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
+
