+++ GitTest/src/main/java/com/smhrd/model/VO/GJ_HS_VO.java
@@ -5,12 +5,12 @@ public class GJ_HS_VO {
 	private int hs_seq;
 	private String hs_nm;
 	private String hs_addr;
-	private int cortarNo;
+	private long cortarNo;
 	private String dong;
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
 	
-	public GJ_HS_VO(int hs_seq, String hs_nm, String hs_addr, int cortarNo, String dong, int lat, int lng) {
+	public GJ_HS_VO(int hs_seq, String hs_nm, String hs_addr, long cortarNo, String dong, double lat, double lng) {
 		this.hs_seq = hs_seq;
 		this.hs_nm = hs_nm;
 		this.hs_addr = hs_addr;
@@ -44,11 +44,11 @@ public class GJ_HS_VO {
 		this.hs_addr = hs_addr;
 	}
 
-	public int getCortarNo() {
+	public long getCortarNo() {
 		return cortarNo;
 	}
 
-	public void setCortarNo(int cortarNo) {
+	public void setCortarNo(long cortarNo) {
 		this.cortarNo = cortarNo;
 	}
 
@@ -60,19 +60,19 @@ public class GJ_HS_VO {
 		this.dong = dong;
 	}
 
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-	public int getLng() {
+	public double getLng() {
 		return lng;
 	}
 
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
