+++ GitTest/src/main/java/com/smhrd/model/VO/GJ_METRO_VO.java
@@ -5,12 +5,13 @@ public class GJ_METRO_VO {
 	private int metro_seq;
 	private String metro_nm;
 	private String metro_addr;
-	private int cortarNo;
+	private long cortarNo;
 	private String dong;
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
 	
-	public GJ_METRO_VO(int metro_seq, String metro_nm, String metro_addr, int cortarNo, String dong, int lat, int lng) {
+	public GJ_METRO_VO(int metro_seq, String metro_nm, String metro_addr, long cortarNo, String dong, double lat,
+			double lng) {
 		this.metro_seq = metro_seq;
 		this.metro_nm = metro_nm;
 		this.metro_addr = metro_addr;
@@ -44,11 +45,11 @@ public class GJ_METRO_VO {
 		this.metro_addr = metro_addr;
 	}
 
-	public int getCortarNo() {
+	public long getCortarNo() {
 		return cortarNo;
 	}
 
-	public void setCortarNo(int cortarNo) {
+	public void setCortarNo(long cortarNo) {
 		this.cortarNo = cortarNo;
 	}
 
@@ -60,22 +61,23 @@ public class GJ_METRO_VO {
 		this.dong = dong;
 	}
 
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-	public int getLng() {
+	public double getLng() {
 		return lng;
 	}
 
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
-
+	
+
