+++ GitTest/src/main/java/com/smhrd/model/VO/GJ_MS_VO.java
@@ -5,12 +5,12 @@ public class GJ_MS_VO {
 	private int ms_seq;
 	private String ms_nm;
 	private String ms_addr;
-	private int cortarNo;
+	private long cortarNo;
 	private String dong;
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
 	
-	public GJ_MS_VO(int ms_seq, String ms_nm, String ms_addr, int cortarNo, String dong, int lat, int lng) {
+	public GJ_MS_VO(int ms_seq, String ms_nm, String ms_addr, long cortarNo, String dong, double lat, double lng) {
 		this.ms_seq = ms_seq;
 		this.ms_nm = ms_nm;
 		this.ms_addr = ms_addr;
@@ -44,11 +44,11 @@ public class GJ_MS_VO {
 		this.ms_addr = ms_addr;
 	}
 
-	public int getCortarNo() {
+	public long getCortarNo() {
 		return cortarNo;
 	}
 
-	public void setCortarNo(int cortarNo) {
+	public void setCortarNo(long cortarNo) {
 		this.cortarNo = cortarNo;
 	}
 
@@ -60,19 +60,19 @@ public class GJ_MS_VO {
 		this.dong = dong;
 	}
 
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-	public int getLng() {
+	public double getLng() {
 		return lng;
 	}
 
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
