+++ GitTest/src/main/java/com/smhrd/model/VO/GJ_PS_VO.java
@@ -5,12 +5,12 @@ public class GJ_PS_VO {
 	private int ps_seq;
 	private String ps_nm;
 	private String ps_addr;
-	private int cortarNo;
+	private long cortarNo;
 	private String dong;
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
 	
-	public GJ_PS_VO(int ps_seq, String ps_nm, String ps_addr, int cortarNo, String dong, int lat, int lng) {
+	public GJ_PS_VO(int ps_seq, String ps_nm, String ps_addr, long cortarNo, String dong, double lat, double lng) {
 		this.ps_seq = ps_seq;
 		this.ps_nm = ps_nm;
 		this.ps_addr = ps_addr;
@@ -44,11 +44,11 @@ public class GJ_PS_VO {
 		this.ps_addr = ps_addr;
 	}
 
-	public int getCortarNo() {
+	public long getCortarNo() {
 		return cortarNo;
 	}
 
-	public void setCortarNo(int cortarNo) {
+	public void setCortarNo(long cortarNo) {
 		this.cortarNo = cortarNo;
 	}
 
@@ -60,22 +60,21 @@ public class GJ_PS_VO {
 		this.dong = dong;
 	}
 
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-	public int getLng() {
+	public double getLng() {
 		return lng;
 	}
 
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
-
