+++ GitTest/src/main/java/com/smhrd/model/VO/GJ_SM_VO.java
@@ -5,12 +5,12 @@ public class GJ_SM_VO {
 	private int sm_seq;
 	private String sm_nm;
 	private String sm_addr;
-	private int cortarNo;
+	private long cortarNo;
 	private String dong;
-	private int lat;
-	private int lng;
+	private double lat;
+	private double lng;
 	
-	public GJ_SM_VO(int sm_seq, String sm_nm, String sm_addr, int cortarNo, String dong, int lat, int lng) {
+	public GJ_SM_VO(int sm_seq, String sm_nm, String sm_addr, long cortarNo, String dong, double lat, double lng) {
 		this.sm_seq = sm_seq;
 		this.sm_nm = sm_nm;
 		this.sm_addr = sm_addr;
@@ -44,11 +44,11 @@ public class GJ_SM_VO {
 		this.sm_addr = sm_addr;
 	}
 
-	public int getCortarNo() {
+	public long getCortarNo() {
 		return cortarNo;
 	}
 
-	public void setCortarNo(int cortarNo) {
+	public void setCortarNo(long cortarNo) {
 		this.cortarNo = cortarNo;
 	}
 
@@ -60,21 +60,23 @@ public class GJ_SM_VO {
 		this.dong = dong;
 	}
 
-	public int getLat() {
+	public double getLat() {
 		return lat;
 	}
 
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 
-	public int getLng() {
+	public double getLng() {
 		return lng;
 	}
 
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
+	
+
